Finds best solution
Uses the Solutions class to find the best solution for a given OCR word
to revert it to the correct word. This is then given back to the
calling class (WordCompareII) in the form of a stack which contains the
best solution to that OCR word to convert it into the correct word.